requirements.txt
Flask==2.0.1
PyMySQL==1.0.2
Flask-Session==0.4
black==21.9b0
python-dotenv==0.19.0
SQLAlchemy==1.4.25
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
wtforms[email]
ipdb==0.13.9
pandas==1.3.4
flask-login==0.5.0
flask-cors==3.0.10
pyOpenSSL==19.0.0
Flask-OAuthlib==0.9.6